Update to next 13.4.16
The latest versions of next and lucide-react do not play nicely with our
jest configuration. So for the time being, we remove the tests that
involved importing an icon, specifically the page.test.tsx file, which
largely existed as an example test file.

This issue was closed, but is still an issue: https://github.com/vercel/next.js/issues/53668