Update app to use edge runtime throughout
Instead of checking for the DATABASE_URL_EDGE, we will now simply assume
edge by default. This has the downside of requiring the use of a service
like Neon for the database, but there are instructions for connecting to
a local postgres like a Neon serverless database.

This has the benefit of assuming everything is edge.

The only `nodejs` explicit endpoint is the next-auth API endpoint, which
is only set to `nodejs` because of limitations with `react-email`. If we
decide to drop email authentication support, this can also be converted
to an edge endpoint.